updating ./Practice-roboshop

+++ components/dispatch.sh
@@ -1,3 +1,11 @@
+
+USER_ID=$(id -u)
+if [ $USER_ID -ne 0]; then
+  echo "You are Non root user"
+  echo "You should run as root user or with sudo to script"
+  exit
+fi
+
 yum install golang -y
 useradd roboshop
 rm -rf dispatch
